content/canvas/test/webgl/non-conf-tests/test_webgl_conformance.html

<!DOCTYPE HTML>
<title>WebGL test: 'webgl' context request</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c"></canvas>
<script>

var canvas = document.getElementById('c');
var gl = canvas.getContext('webgl');

var isMobile = /Mobile/.test(navigator.userAgent);
var isTablet = /Tablet/.test(navigator.userAgent);
var shouldBeConformant = !isMobile && !isTablet;

if (shouldBeConformant) {
  // Desktop.
  ok(gl, 'Expected conformance on: ' + navigator.userAgent);
} else {
  // Fennec, B2G.
  ok(!gl, 'Expected no conformance on: ' + navigator.userAgent);
}

</script>
